CPB-760: Replace LocalStack TestContainer with Floci
